requirements.txt
numpy
torch
ujson
tqdm
transformers==4.30.0
opt_einsum
IPython
iterative-stratification
psutil
maverick-coref
